Various improvements
* Improved postinst testing procedure, added guard against root
* Added dhcp-client options to supersede domain name and dns
* Added mtu 9000 to backbone interface
* Introduced rc.sci with local post-boot tunings
* Added blockdev tuning for /dev/sd*
* Default vnc_cluster_password if none configured
* Automatic RAID10 prepare for xenvg with ability to chose far/near
* Automatic prepare of LVM/xenvg/system-stuff if none
* cdimages and ganeti export are going to /stuff -> LVM/xenvg/system-stuff

+++ profiles/SCI-amd64.files/postinst.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 # this is the post-install script
 # the newly-installed system is not yet booted, but chrooted at the moment of execution
-# XXX xend-relocation-hosts is turned off (can be tuned later via puppet)
 
 set -x
 
@@ -19,10 +18,14 @@ fi
 if [ "$1" = "real" ]; then
  target=""
 else # prepare test environment
+ if [ `id -u` -eq 0 ]; then
+   echo "Please don't run test mode as root, because it can modify your system accidentally!"
+   exit 1
+ fi
  rm -rf target
  cp -a target-orig target
  target=target
- mkdir -p $target/etc/xen $target/usr/sbin $target/usr/share/ganeti $target/usr/lib/xen
+ mkdir -p $target/etc/xen $target/usr/sbin $target/usr/share/ganeti $target/usr/lib/xen $target/usr/local/sbin
 fi
 
 cp -a real/* .
@@ -37,7 +40,8 @@ for i in \
  /etc/default/puppet \
  /etc/default/xendomains \
  /etc/modules \
- /etc/rsyslog.conf
+ /etc/rsyslog.conf \
+ /etc/dhcp/dhclient.conf
 do
  cp $target/$i backup
 done
@@ -87,7 +91,12 @@ else
 fi
 fi
 
+## Assign supersede parameters for node's dhcp
+dns=`grep nameserver $target/etc/resolv.conf|awk '{print $2; exit}'`
+./strreplace.sh $target/etc/dhcp/dhclient.conf "^#supersede domain-name" "supersede domain-name $domain\nsupersede domain-name-servers $dns"
+
 ## Set default interface to be bridged, optionally with vlan (see postinst.conf)
+## Set mtu 9000 on the default interface
 
 echo Configuring interfaces
 ifs=$target/etc/network/interfaces
@@ -128,6 +137,8 @@ cat <<EOF >>interfaces
         bridge_ports $xenif
         bridge_stp off
         bridge_fd 0
+	up ifconfig $xenif mtu 9000
+	up ifconfig xen-br0 mtu 9000
 EOF
  
 ## Add example of additional interfaces
@@ -182,10 +193,153 @@ echo Setting up defaults
 
 ./strreplace.sh $target/etc/default/xendomains "^XENDOMAINS_SAVE" 'XENDOMAINS_SAVE=""'
 
-## Set up CD-ROM repository: create /var/lib/cdimages, /media/sci
+## Add startup script rc.sci to setup performance
+
+# a bit ugly, but fast ;)
+cat <<EOF >$target/etc/rc.local
+#!/bin/sh -e
+#
+# rc.local
+#
+# This script is executed at the end of each multiuser runlevel.
+# Make sure that the script will "exit 0" on success or any other
+# value on error.
+#
+# In order to enable or disable this script just change the execution
+# bits.
+
+if [ -f /etc/rc.sci ]; then
+  . /etc/rc.sci
+fi
+
+exit 0
+EOF
+chmod +x $target/etc/rc.local
+
+## Add "xm sched-credit -d0 -w512" to rc.sci
+# equal priority of Dom0 make problems on the block devices
+
+## Tune storage scheduler for better disk latency
+cat <<EOFF >$target/etc/rc.sci
+#!/bin/sh
+# On-boot configuration for hardware for better cluster performance
+# mostly http://code.google.com/p/ganeti/wiki/PerformanceTuning
+
+# rise priority for dom0, alowing drbd to work fine
+xm sched-credit -d0 -w512
+
+modprobe sg
+disks=\`sg_map -i|awk '{if(\$3=="ATA"){print substr(\$2, length(\$2))}}'\`
+for i in \$disks; do
+  # Set value if you want to use read-ahead
+  ra="$read_ahead"
+  if [ -n "\$ra" ]; then
+    blockdev --setra \$ra /dev/sd\$i
+  fi
+  if grep -q sd\$i /etc/sysfs.conf; then
+    echo sd\$i already configured in /etc/sysfs.conf
+  else
+ cat <<EOF >>/etc/sysfs.conf
+block/sd\$i/queue/scheduler = deadline
+block/sd\$i/queue/iosched/front_merges = 0
+block/sd\$i/queue/iosched/read_expire = 150
+block/sd\$i/queue/iosched/write_expire = 1500
+EOF
+  fi
+done
+/etc/init.d/sysfsutils restart
+EOFF
+chmod +x $target/etc/rc.sci
+
+## Add tcp buffers tuning for drbd
+## Tune disk system to avoid (or reduce?) deadlocks
+cat <<EOF >$target/etc/sysctl.d/sci.conf
+# Increase "minimum" (and default) 
+# tcp buffer to increase the chance to make progress in IO via tcp, 
+# even under memory pressure. 
+# These numbers need to be confirmed - probably a bad example.
+#net.ipv4.tcp_rmem = 131072 131072 10485760 
+#net.ipv4.tcp_wmem = 131072 131072 10485760 
+
+# add disk tuning options to avoid (or reduce?) deadlocks
+# gives better latency on heavy load
+vm.swappiness = 0
+vm.overcommit_memory = 1
+vm.dirty_background_ratio = 5
+vm.dirty_ratio = 10
+vm.dirty_expire_centisecs = 1000
+EOF
+
+## Add workaround for bnx2x NIC on HP Proliant and Blade servers
+# https://bugzilla.redhat.com/show_bug.cgi?id=518531
+echo "options bnx2x disable_tpa=1" >$target/etc/modprobe.d/sci.conf
+
+## Set up symlinks /boot/vmlinuz-2.6-xenU, /boot/initrd-2.6-xenU
+
+# we'll assume only one xen kernel at the moment of the installation
+ln -s $target/boot/vmlinuz-2.6.*-xen-amd64 $target/boot/vmlinuz-2.6-xenU
+ln -s $target/boot/initrd.img-2.6.*-xen-amd64 $target/boot/initrd.img-2.6-xenU
+
+## Set up symlink /usr/lib/xen for quemu-dm (workaround)
+ln -s $target/usr/lib/xen-4.0 $target/usr/lib/xen
+
+## Set vnc master password if provided in postinst.conf
+echo "${vnc_cluster_password:=gntwin}" >$target/etc/ganeti/vnc-cluster-password
+chmod 600 $target/etc/ganeti/vnc-cluster-password
+
+if [ ! -f /proc/mounts ]; then
+	echo Warning: /proc is not mounted. Trying to fix.
+	mkdir -p /proc
+	mount /proc
+	proc_mounted=1
+fi
+
+## Create RAID10 with far layout and LVM/xenvg
+# if there is xenvg_disks and xenvg_md then try to create RAID and LVM
+# else expect the xenvg is already configured
+if [ -n "$xenvg_disks" -a -n "$xenvg_md" ]; then
+  echo ...Creating RAID10 with far layout
+  # No preexisting checks - it should simply fail the already completed phases
+  ndisks=`ls $xenvg_disks|wc -w`
+  if [ -n "$xenvg_spares" ]; then
+    spares="`ls $xenvg_spares|wc -w` $xenvg_spares"
+  fi
+  echo "CALLING: mdadm --create -l 10 -n $ndisks --layout=${md_layout:-n2} $xenvg_md $xenvg_disks $spares"
+  mdadm --create -l 10 -n $ndisks --layout=${md_layout:-n2} $xenvg_md $xenvg_disks $spares
+  /usr/share/mdadm/mkconf >$target/etc/mdadm/mdadm.conf
+  # XXX is it needed? will -u reflect right kernel? or better will be -a?
+  #update-initramfs -u
+  vgcreate xenvg $xenvg_md
+fi
+
+## Prepare LVM/xenvg/system-stuff on /stuff for cd images, dumps etc.
+# if /stuff is already present, let's suppose that it is fully configured
+if [ ! -d $target/stuff ]; then
+  echo ...Creating /stuff volume
+  mkdir $target/stuff
+  # don't touch data if volume exists
+  if [ ! -b /dev/xenvg/system-stuff ]; then
+    echo "CALLING: lvcreate -v -L ${stuff_volume_size:-20G} -n system-stuff xenvg"
+    lvcreate -v --noudevsync -L ${stuff_volume_size:-20G} -n system-stuff xenvg
+    if [ $? -eq 0 ]; then
+      sleep 1 # XXX for a case
+      if [ -b /dev/xenvg/system-stuff ]; then
+	echo ...Formatting new xenvg/system-stuff
+	mkfs.ext4 /dev/xenvg/system-stuff
+      fi
+    fi
+  fi
+  # recheck volume and mount
+  if [ -b /dev/xenvg/system-stuff ]; then
+    echo "/dev/xenvg/system-stuff /stuff ext4 errors=remount-ro 0 0" >>$target/etc/fstab
+    mount /stuff
+  fi
+fi
+
+## Set up CD-ROM repository: create /stuff/cdimages, /media/sci
 
 echo Setting up local CD-ROM repository
-mkdir -p $target/var/lib/cdimages
+mkdir -p $target/stuff/cdimages
 mkdir -p $target/media/sci
 
 cat <<EOF >>$target/etc/apt/apt.conf.d/99-sci
@@ -202,65 +356,7 @@ SUITE=squeeze
 EXTRA_PKGS="linux-image-xen-amd64"
 EOF
 
-## Add "xm sched-credit -d0 -w512" to /etc/rc.local
-# equal priority of Dom0 make problems on the block devices
-## Add sysfs tuning for better disk latency and to avoid kernel problems
-
-cat <<EOF >$target/etc/rc.local
-#!/bin/sh -e
-#
-# rc.local
-#
-# This script is executed at the end of each multiuser runlevel.
-# Make sure that the script will "exit 0" on success or any other
-# value on error.
-#
-# In order to enable or disable this script just change the execution
-# bits.
-
-# rise priority for dom0, alowing drbd to work fine
-xm sched-credit -d0 -w512
-
-# add disk tuning options to avoid (or reduce?) deadlocks
-# gives better latency on heavy load
-echo "0" >/proc/sys/vm/swappiness
-echo "1" >/proc/sys/vm/overcommit_memory
-echo "5" >/proc/sys/vm/dirty_background_ratio
-echo "10" >/proc/sys/vm/dirty_ratio
-echo "1000" >/proc/sys/vm/dirty_expire_centisecs
-
-exit 0
-EOF
-
-chmod +x $target/etc/rc.local
-
-## Add workaround for bnx2x NIC on HP Proliant and Blade servers
-# https://bugzilla.redhat.com/show_bug.cgi?id=518531
-echo "options bnx2x disable_tpa=1" >$target/etc/modprobe.d/sci.conf
-
-## Set up symlinks /boot/vmlinuz-2.6-xenU, /boot/initrd-2.6-xenU
-
-# we'll assume only one xen kernel at the moment of the installation
-ln -s $target/boot/vmlinuz-2.6.*-xen-amd64 $target/boot/vmlinuz-2.6-xenU
-ln -s $target/boot/initrd.img-2.6.*-xen-amd64 $target/boot/initrd.img-2.6-xenU
-
-## Set up symlink /usr/lib/xen for quemu-dm (workaround)
-ln -s $target/usr/lib/xen-4.0 $target/usr/lib/xen
-
-## Set vnc master password if provided in postinst.conf
-if [ -n "$vnc_cluster_password" ]; then
-	echo "$vnc_cluster_password" >$target/etc/ganeti/vnc-cluster-password
-	chmod 600 $target/etc/ganeti/vnc-cluster-password
-fi
-
-if [ ! -f /proc/mounts ]; then
-	echo Warning: /proc is not mounted. Trying to fix.
-	mkdir -p /proc
-	mount /proc
-	proc_mounted=1
-fi
-
-## Copy-in SCI-CD iso image to /var/lib/cdimages, mount to /media/sci, set up sources.list
+## Copy-in SCI-CD iso image to /stuff/cdimages, mount to /media/sci, set up sources.list
 
 # when installing from USB stick, two /cdrom mounts are shown
 # XXX 'head -1' may be a wrong choice here, but will not differ them at present
@@ -268,18 +364,29 @@ dev=`grep '/cdrom' /proc/mounts|head -1|cut -d' ' -f1`
 
 if [ -n "$dev" -a -e "$dev" ]; then
 	echo ...Copying CD-ROM image
-	dd if=$dev of=$target/var/lib/cdimages/sci.iso
+	dd if=$dev of=$target/stuff/cdimages/sci.iso
 
-	echo "/var/lib/cdimages/sci.iso /media/sci iso9660 loop 0 0" >>$target/etc/fstab
+	echo "/stuff/cdimages/sci.iso /media/sci iso9660 loop 0 1" >>$target/etc/fstab
 
 	echo ...Adding repository data
 	mount /media/sci && (apt-cdrom -d=/media/sci add; umount /media/sci)
 else
 	echo Unable to find CD-ROM device
-	echo "#/var/lib/cdimages/sci.iso /media/sci iso9660 loop 0 0" >>$target/etc/fstab
+	echo "#/stuff/cdimages/sci.iso /media/sci iso9660 loop 0 1" >>$target/etc/fstab
 fi
 
 test -n "$proc_mounted" && umount /proc
+umount /stuff
+
+## Link /var/lib/ganeti/export to /stuff/export
+
+mkdir $target/stuff/export
+(cd $target/var/lib/ganeti && ln -s $target/stuff/export)
+
+## Patch ganeti for viridian option
+
+source=`pwd`
+(cd $target/usr/share/pyshared/ganeti; patch -p1 <$source/patch/ganeti-2.5.0-viridian.patch)
 
 ## Add ganeti hooks if any
 
@@ -292,17 +399,22 @@ mkdir -p $target/etc/ganeti/instance-debootstrap/hooks
 cp -r files/ganeti/instance-debootstrap/hooks/* $target/etc/ganeti/instance-debootstrap/hooks/
 
 ## Add ganeti-instance-debootstrap variant "sci"
+
 mkdir -p $target/etc/ganeti/instance-debootstrap/variants
 cp -r files/ganeti/instance-debootstrap/variants/* $target/etc/ganeti/instance-debootstrap/variants/
 echo sci >>$target/etc/ganeti/instance-debootstrap/variants.list
 
-## Add ganeti OS "windows" scripts (ntfsclone - based)
+## Add ganeti OS "windows" scripts (ntfsclone - based) and simple "raw"
+
 cp -r files/os $target/usr/share/ganeti/
 
 ## Add SCI deploing scripts
+## Make LV names 'system-.*' ignored by ganeti
+
 cp files/sbin/* $target/usr/local/sbin/
 
 ## Filling SCI configuration template
+
 mkdir $target/etc/sci
 cat <<EOF >$target/etc/sci/sci.conf
 # The cluster's name and IP. They MUST be different from node names
